initial move of details component to beta folder
