+++ app/components/primer/beta/details.html.erb
