+++ app/components/primer/dropdown.html.erb
@@ -1,4 +1,4 @@
-<%= render(Primer::DetailsComponent.new(**@system_arguments)) do |c| %>
+<%= render(Primer::Beta::Details.new(**@system_arguments)) do |c| %>
   <% c.summary(**@button_arguments) do %>
     <%= button %>
   <% end %>
