+++ lib/rubocop/cop/primer/component_name_migration.rb
@@ -15,6 +15,7 @@ module RuboCop
       # Primer::Beta::ComponentName.new()
       class ComponentNameMigration < BaseCop
         DEPRECATIONS = {
+          "Primer::DetailsComponent" => "Primer::Beta::Details",
           "Primer::BoxComponent" => "Primer::Box",
           "Primer::BlankslateComponent" => "Primer::Beta::Blankslate",
           "Primer::BorderBoxComponent" => "Primer::Beta::BorderBox",
