+++ static/statuses.json
@@ -22,6 +22,7 @@
   "Primer::Beta::BorderBox::Header": "alpha",
   "Primer::Beta::Breadcrumbs": "beta",
   "Primer::Beta::Breadcrumbs::Item": "alpha",
+  "Primer::Beta::Details": "beta",
   "Primer::Beta::Flash": "beta",
   "Primer::Beta::Text": "beta",
   "Primer::Beta::Truncate": "beta",
@@ -37,7 +38,7 @@
   "Primer::ConditionalWrapper": "alpha",
   "Primer::Content": "stable",
   "Primer::CounterComponent": "beta",
-  "Primer::DetailsComponent": "beta",
+  "Primer::DetailsComponent": "deprecated",
   "Primer::Dropdown": "alpha",
   "Primer::Dropdown::Menu": "alpha",
   "Primer::Dropdown::Menu::Item": "alpha",
