+++ test/components/beta/details_test.rb
@@ -2,11 +2,11 @@
 
 require "test_helper"
 
-class PrimerDetailsComponentTest < Minitest::Test
+class PrimerBetaDetailsTest < Minitest::Test
   include Primer::ComponentTestHelpers
 
   def test_overlay_default_renders_details_overlay
-    render_inline(Primer::DetailsComponent.new(overlay: :default)) do |component|
+    render_inline(Primer::Beta::Details.new(overlay: :default)) do |component|
       component.summary do
         "Summary"
       end
@@ -19,7 +19,7 @@ class PrimerDetailsComponentTest < Minitest::Test
   end
 
   def test_overlay_dark_renders_details_overlay_dark
-    render_inline(Primer::DetailsComponent.new(overlay: :dark)) do |component|
+    render_inline(Primer::Beta::Details.new(overlay: :dark)) do |component|
       component.summary do
         "Summary"
       end
@@ -32,7 +32,7 @@ class PrimerDetailsComponentTest < Minitest::Test
   end
 
   def test_renders_details_reset_when_resetting_the_button_style
-    render_inline(Primer::DetailsComponent.new(reset: true)) do |component|
+    render_inline(Primer::Beta::Details.new(reset: true)) do |component|
       component.summary do
         "Summary"
       end
@@ -45,7 +45,7 @@ class PrimerDetailsComponentTest < Minitest::Test
   end
 
   def test_default_component_renders_btn_summary
-    render_inline(Primer::DetailsComponent.new) do |component|
+    render_inline(Primer::Beta::Details.new) do |component|
       component.summary do
         "Summary"
       end
@@ -58,7 +58,7 @@ class PrimerDetailsComponentTest < Minitest::Test
   end
 
   def test_does_not_render_btn_if_button_false
-    render_inline(Primer::DetailsComponent.new) do |component|
+    render_inline(Primer::Beta::Details.new) do |component|
       component.summary(button: false) do
         "Summary"
       end
@@ -73,7 +73,7 @@ class PrimerDetailsComponentTest < Minitest::Test
 
   def test_falls_back_to_defaults_when_invalid_overlay_is_passed
     without_fetch_or_fallback_raises do
-      render_inline(Primer::DetailsComponent.new(overlay: :bar)) do |component|
+      render_inline(Primer::Beta::Details.new(overlay: :bar)) do |component|
         component.summary { "Summary" }
         component.body { "Body" }
       end
@@ -85,7 +85,7 @@ class PrimerDetailsComponentTest < Minitest::Test
 
   def test_falls_back_to_default_body_tag_when_invalid_body_tag_is_passed
     without_fetch_or_fallback_raises do
-      render_inline(Primer::DetailsComponent.new) do |component|
+      render_inline(Primer::Beta::Details.new) do |component|
         component.summary { "Summary" }
         component.body(tag: :img) { "Body" }
       end
@@ -96,7 +96,7 @@ class PrimerDetailsComponentTest < Minitest::Test
   end
 
   def test_passes_props_to_button
-    render_inline(Primer::DetailsComponent.new) do |component|
+    render_inline(Primer::Beta::Details.new) do |component|
       component.summary(size: :small, scheme: :primary) do
         "Summary"
       end
@@ -109,11 +109,11 @@ class PrimerDetailsComponentTest < Minitest::Test
   end
 
   def test_prevents_rendering_without_slots
-    render_inline(Primer::DetailsComponent.new)
-    render_inline(Primer::DetailsComponent.new) do |component|
+    render_inline(Primer::Beta::Details.new)
+    render_inline(Primer::Beta::Details.new) do |component|
       component.body { "Body" }
     end
-    render_inline(Primer::DetailsComponent.new) do |component|
+    render_inline(Primer::Beta::Details.new) do |component|
       component.summary { "Summary" }
     end
 
@@ -122,6 +122,6 @@ class PrimerDetailsComponentTest < Minitest::Test
   end
 
   def test_status
-    assert_component_state(Primer::DetailsComponent, :beta)
+    assert_component_state(Primer::Beta::Details, :beta)
   end
 end
